+++ client/lib/js/functions.js
@@ -58,12 +58,21 @@
  * parameters to return.
  * This function doesn't supports [lazy evaluation](https://en.wikipedia.org/wiki/Lazy_evaluation).
  *
+ * - `get` - returns the numerical N-value of a 0-base array.
+ *
+ * - `slice` - returns a subarray starting from `start` to `end-1` of a 0-base array.
+ *
  * ## Examples
  *
  * @example = substr('ABeamer', 4, 4)
  * @example = round(12.4)
+ * @example = get([10,30,15],1)
+ * it returns `30`
+ * @example = slice([10,30,15,40],1, 3)
+ * it returns `[30, 15]`
  *
  * ## Arrays
+ *
  * Since ABeamer 1.6 that single numerical argument functions also support arrays.
  * The operation is perform for each element, and it returns an array.
  * The array functions can be composed.
@@ -181,6 +190,27 @@ var ABeamer;
         req.res.paType = res.paType;
         req.res.sValue = res.sValue;
         req.res.numValue = res.numValue;
+    };
+    ABeamer._exFunctions['get'] = function (params, req) {
+        req.checkParams(req, 2, [3 /* Array */, 1 /* Number */]);
+        var index = Math.round(params[1].numValue);
+        var arr = params[0].arrayValue;
+        if (index < 0 || index >= arr.length) {
+            ABeamer.throwErr('Invalid indexing');
+        }
+        req.res.paType = 1 /* Number */;
+        req.res.numValue = arr[index];
+    };
+    ABeamer._exFunctions['slice'] = function (params, req) {
+        req.checkParams(req, 3, [3 /* Array */, 1 /* Number */, 1 /* Number */]);
+        var start = Math.round(params[1].numValue);
+        var end = Math.round(params[2].numValue);
+        var arr = params[0].arrayValue;
+        if (start < 0 || start >= arr.length || end < 0 || end >= arr.length) {
+            ABeamer.throwErr('Invalid indexing');
+        }
+        req.res.paType = 3 /* Array */;
+        req.res.arrayValue = arr.slice(start, end);
     };
 })(ABeamer || (ABeamer = {}));
 //# sourceMappingURL=functions.js.map