+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   > With this addition, those functions can process individual elements inside an array,
   > and set the result to an array.
 - Add new `statistic-functions` plugin with 8 fundamental functions.
+- Add `get` and `slice` functions for array data extraction to `functions`.
 
 
 ## [1.5.0] -